전투 - 슬라임 공격 코드 수정 - 아이템 합치기, HUDUI 갱신, 애니메이션

+++ Assets/ProjectSV/Scripts/BreakbleObject.cs
@@ -13,7 +13,7 @@ public class BreakbleObject : MonoBehaviour, IDamage
         currentHP = maxHP;
     }
 
-    public void TakeDamage(IActor actor, int damage)
+    public void TakeDamage(IActor actor, float damage)
     {
         // 애니메이터
         // 이펙트
