+++ Assets/ProjectSV/Scripts/PlayerCharacter/CharacterAttribute.cs
@@ -11,8 +11,8 @@ public class CharacterAttribute : MonoBehaviour
     public float ModifierValue { get; set; } // 영구/임시, 증가/하락, 최대값 변화값 (레벨, 장비, 아이템, 스킬, ...)
     public float BuffedValue { get; set; } // 임시, 증가/하락, 현재값 변화값 (체력 제외 맥스 넘어서기 가능)
 
-    public System.Action<int, int> OnModifierChanged; // 최대, 현재
-    public System.Action<int, int> OnBuffedChanged; // 최대, 현재
+    public System.Action<float, float> OnModifierChanged; // 최대, 현재
+    public System.Action<float, float> OnBuffedChanged; // 최대, 현재
 }
 
 // M, B 헷갈리니까 실제 수치 바꾸면서 구분하기
